+++ feincms3_sites/__init__.py
@@ -1,4 +1,4 @@
-VERSION = (0, 1, 1)
+VERSION = (0, 2, 0)
 __version__ = '.'.join(map(str, VERSION))
 
 default_app_config = 'feincms3_sites.apps.SitesAppConfig'