+++ CHANGELOG.rst
@@ -5,6 +5,9 @@ Change log
 `Next version`_
 ~~~~~~~~~~~~~~~
 
+`0.2`_ (2018-04-17)
+~~~~~~~~~~~~~~~~~~~
+
 - Added documentation (README only for now).
 - Made ``AppsMiddleware`` raise a 404 error if no site matches the
   current requests' host.
@@ -29,4 +32,4 @@ Change log
 
 .. _0.1: https://github.com/matthiask/feincms3-sites/commit/e19c1ebef0
 .. _0.2: https://github.com/matthiask/feincms3-sites/compare/0.1...0.2
-.. _Next version: https://github.com/matthiask/feincms3-sites/compare/0.1...master
+.. _Next version: https://github.com/matthiask/feincms3-sites/compare/0.2...master